refactor!: separate hashing functions into new subproject (#593)
